Remove ting_mark_log and ting_marc.client all together
This function doesn't seem to be used anywhere and it's wrongly named. Since this was the last function there, ting_marc.client.inc was removed.
If logging should be re-added it would probably make more sense to add it in the module file. Otherwise, this PR can be reverted to bring it back.